+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -59,10 +59,7 @@ import at.hannibal2.skyhanni.features.mobs.AshfangMinisNametagHider
 import at.hannibal2.skyhanni.features.mobs.MobHighlight
 import at.hannibal2.skyhanni.features.nether.ashfang.*
 import at.hannibal2.skyhanni.features.nether.reputationhelper.CrimsonIsleReputationHelper
-import at.hannibal2.skyhanni.features.rift.CruxWarnings
-import at.hannibal2.skyhanni.features.rift.HighlightRiftGuide
-import at.hannibal2.skyhanni.features.rift.RiftAPI
-import at.hannibal2.skyhanni.features.rift.RiftTimer
+import at.hannibal2.skyhanni.features.rift.*
 import at.hannibal2.skyhanni.features.slayer.*
 import at.hannibal2.skyhanni.features.slayer.blaze.BlazeSlayerClearView
 import at.hannibal2.skyhanni.features.slayer.blaze.BlazeSlayerDaggerHelper
@@ -308,6 +305,7 @@ class SkyHanniMod {
         loadModule(RiftTimer())
         loadModule(HighlightRiftGuide())
         loadModule(CruxWarnings())
+        loadModule(RiftLarva())
 
         init()